CASMPET-4105: Update cray-precache-images for new istio and unbound
CASMPET-4080 updated istio to 1.7.8 and that brought in a new
proxyv2 image. The cray-precache-images list should have been
changed to the new image (1.7.8-cray1).

Also I noticed that the cray-precache-images pods on a system were
reporting that the cray-dns-unbound images also couldn't be found
so I updated that image to match what's in the docker/index.html
file.