CASMPET-4147: Update cray-precache-images to 0.3.0
This picks up cray-precache-images to pick up:

CASMPET-4095: The cray-precache-images pods should not have istio sidecar

Also, I added a comment to images that are in the manifest to
be pre-cached so that developers remember to keep it up to date.

I also noticed that cray-dns-unbound was out of date with the
docker index so updated that.