CASMUSER-2805 update broker UAI, basic UAI and update-uas to 1.2.4
The Broker UAI and Basic UAI have been moved together into
the uai-images repo, along with the update-uas job and Helm
chart.  This permits update-uas to register both the broker
and basic images with the correct version in UAS and to
update classes / default image settings to the specified
version of the image automatically. The first version of all
of this that can be managed this way is 1.2.4.