CASMPET-5000: Updates for new cray-kiali chart
CASMPET-3926 changes the istio configuration to not deploy Kiali.
Kiali is now deployed using the new cray-kiali chart.

I put the deployment of the new cray-kiali chart right after
cray-istio because Kiali deployment was done by cray-istio but
it could go anywhere after cray-istio.

The new Kiali images are built by the container-images repo and
published to algol60.

This change is targeted to CSM-1.2.

That's the only change that was released for CASMPET-5000.