CASMPET-5158: Update keycloak charts for keycloak-setup image
This includes the following changes:

* CASMPET-4299: Configure new nexus Keycloak client and K8s secret
* CASMPET-4714: keycloak-setup create oauth2-proxy client and Secret

This also requires a change to the customizations.yaml to pass the
proxiedHosts list for the new oauth2-proxy client to the
cray-keycloak chart.

The proxied hosts list in cray-keycloak was referenced from the
cray-keycloak-gatekeeper config. Since the cray-keycloak-gatekeeper
is going away soon, I pulled the list from cray-keycloak-gatekeeper
to a new variable in the `spec` section.