MTL-1613 New Metal-iPXE with DHCP Optimization
This removes the "DHCP on every NIC" patch that workedaround
CASMINST-3789, thus restoring the optimization and reducing the
touching of the MLAG. CASMINST-3789 is fixed a different way in the
source code of metal-ipxe while keeping the performance optimization.
Specifically a bad variable was being used that previously worked fine
until CASMTRIAGE-2789 changes went in, the variable in question became
ambiguous and ceased reliably providing the xname.