CASMINST-4968: ignore errors from create-kis-artifacts.sh for now
For some reason, create-kis-artifacts.sh is unable to unmount a
bind mount within an NCN chroot environment, but *is* able to do
so in the PIT environment. This needs further investigation. The
change in this commit allows it to compensate for the error on
NCNs, while being a no-op on the PIT.