CASMINST-5003
This brings in a new pit-init which adds `CSM_RELEASE` to the boot
parameters for the NCNs, images will download into a directory that
corresponds to the `CSM_RELEASE`. This aligns the install path with the
recent changes (CASMINST-4976) in the upgrade path for supporting the
auto-wipe on storage nodes (while providing a feature enhancement to k8s
nodes).