CASMCMS-8239: Lengthen BOS deployment timeout to 10 minutes
When BOS V2 performs an upgrade, it needs to migrate data from the etcd
database to a new redis database. It does this migration in a
post-upgrade Helm hook. The Helm hook needs to wait for the pod
containing the etcd database to start up before it can do its migration
work. The etcd pod can take a long time to start up and this causes
Helm to time out the deployment. Extending the timeout avoids this
failure.