CASMTRIAGE-4693 Workaround CASMPET-6205
The cray-ceph-rbd pods use an older version of the kmod package that is
incompatible with the new kernel's Zstandard compressed modules. To work
around this, we leverage the host's kmod package and invoke `modprobe`
directly on the host. By loading the module via the host the pods never
attempt to run `modprobe` and run peacefully.