MTL-2076 New `libcsm`
The new `libCSM` uses a `virtualenv`, no longer installing into the
system Python environment. The package still depends on `python39-base`
or `python310-base` (for SP3 and SP4 respectively) to be installed.

This entailed a package name change since it is now all-inclusive, the
canonical naming of `pythonXY-<name>` is only applicable to packages
installing into the system Python environment.