MTL-2000 Add `cloud-init.yaml`
This stages a new manifest for packages to install via cloud-init. The
new `cloud-init.yaml` file will be read during fresh installs and
upgrades by `pit-init.sh` and `prerequisites.sh` in order to generate
cloud-init data.

This is just staging the change, eventually this will replace
`lib/install-goss-tests.sh` once the change is made active
through a new NCN image once everything is staged.

A new `cray-site-init` and `pit-init` are pulled in that has the ability to read this
file, but this is also a staging change.