CRAYSAT-1741: Update cray-sat version
Update the version of the cray-sat container image. This will pull in
support for specifying the new DKMS parameter in layers of CFS
configurations created by `sat bootprep`.

Test Description:
This new version of the cray-sat container image was tested as described
in the commit and pull request to the sat repository.