noos iuf-cli and new libcsm
MTL-2000 is failing to find iuf-cli and libcsm in sle-15sp5 and noos.

This pulls in a new iuf-cli that publishes to noos, as well as a new
libcsm that publishes to sp3/sp4/sp5.

This also copies loftsman into the sp5 index.yaml file for sp5, as it
also publishes for 3/4/5 (it should be switched to noos soon).

(cherry picked from commit d2f5916d8ab3815caf913d729110b1ed299182e5)