noos iuf-cli; new libcsm and csi
MTL-2000 is failing to find iuf-cli and libcsm in sle-15sp5 and noos.

This pulls in a new iuf-cli that publishes to noos, as well as a new
libcsm that publishes to sp3/sp4/sp5.

Additionally CSI was tested again with MTL-2000 and needed some minor
tweaks to how it patches the data structure.

This also copies loftsman into the sp5 index.yaml file for sp5, as it
also publishes for 3/4/5 (it should be switched to noos soon).