CASMINST-6597: Add update-mgmt-ncn-cfs-config.sh script
This script is used to update only the CSM-provided layers of a CFS
configuration. This is useful during a patch installation when `sat
bootprep` may not be run with a whole set of new products.

Documentation will be added to the patch procedure to describe how to
use this script.

This script was mostly copied from the sat-product-stream repository.
It uses existing shell functions defined in the hpc-shastarelm-release
shared library which is already git-vendor'd into this repository. Those
functions run the `cfs-config-util` container image with `podman`.

Use version 5.0.0 of cfs-config-util. Eventually, it would be good to
update the default version in hpc-shastarelm-release, but since it's a
backwards-incompatible major version change, we would need to announce
that and help teams migrate.

Test Description:
Built a stripped down CSM release distribution tar file, copied it to
mug, modified the product catalog to look like the matching version of
CSM had been installed and provided a new branch of the
csm-config-management repo in VCS, and then ran this script with the
following arguments:

```
./update-mgmt-ncn-cfs-config.sh --base-query role=management --save \
    --create-backups
```

Verified that this updated the CFS configuration which was currently set
as the desiredConfig on the CFS components corresponding to the
management nodes and then waited for CFS batcher to configure them. Also
verified that it created an update of the configuration before the
modifications were made.

(cherry picked from commit 64ef3d60163453e16ad54f6bfb35c8619f12d544)
(cherry picked from commit e362a21a374e7e0d5c8de10b47e42bbc5e251d20)