CRAYSAT-1774: Add cray-sat-podman RPM to noos repository
This package was part of the SAT product stream and is being added to
the CSM product stream to merge the SAT product into CSM fully.

Technically this package is also already installed in the Kubernetes
management NCN images, but it seems like a good idea to keep including
it in a Nexus package repository like SAT does today. This would enable
updates of this package if needed without updating the management NCN
images.

Test Description:
The Jenkins build should validate that the RPM can be downloaded from
the location specified in the `index.yaml` manifest file.