CRAYSAT-1767: Set cray-sat version in csm-config
Add an initContainer definition to the values defined for the
`csm-config` Helm chart. This initContainer sets the cray-sat image tag
to match the version included in the CSM build, so that they are always
consistent. The Ansible content in csm-config sets this cray-sat
container image tag in the file `/opt/cray/etc/sat/version`, so that the
sat-podman wrapper script knows which version of the image to run.

Add logic to `release.sh` that inspects the container image tags
specified in `docker/index.yaml` and injects those tags for the cray-sat
container image and the Alpine Linux container image into the csm-config
Helm chart values in the Loftsman manifest.

Modify `lib/setup-nexus.sh` to set the `cray-sat` container image
version and tag the latest one as `csm-latest`. We are phasing out
reliance on this `csm-latest` tag, but it is still the default used
by the sat-podman wrapper script if the file `/opt/cray/etc/sat/version`
does not exist. Modify `lib/setup-nexus.sh` to also write the tag of the
`cray-sat` image to `/opt/cray/etc/sat/version` as well. This is only
executed on `ncn-m001`, but that is sufficient for bootstrapping. When
the CSM layer of the CFS configuration is applied, this file will be set
on all master and worker management nodes.

Test Description:
Only tested one piece at a time so far.