CRAYSAT-1767: Ensure `/opt/cray/etc/sat` dir exists in setup-nexus.sh
Update `lib/setup-nexus.sh` to ensure the directory `/opt/cray/etc/sat`
exists before writing to the `version` file in that directory. On an
upgrade, this directory should already exist because it will have been
created by management node personalization performed by CFS. However, on
a fresh install, this script is executed on the PIT, and this directory
will not exist. In fact, in that case, `sat` is not even installed on
the PIT, so writing to this file won't affect anything, but it also
won't hurt anything.