CASMINST-6697: Modify `update-mgmt-ncn-cfs-config.sh` for CSM 1.3
There is a slight difference in CSM 1.3 compared to CSM 1.4 in which CSM
layers will be present in the CFS configuration which is applied to
management nodes. Specifically, since the Papaya release (CSM 1.3) did
not yet combine the two CFS configurations used for management node
image customization and node personalization into a single CFS
configuration, only the `site.yml` Ansible playbook was used in the CFS
configuration applied via node personalization. The `ncn-initrd.yml`
playbook was only used in a separate CFS configuration specifically for
customization of Kubernetes worker management nodes.