CRAYSAT-1767: Fix sysmgmt.yaml loftsman manifest
Loftsman or Helm wants the second arg value to be fully quoted. I think
it is because it contains a colon, and it interprets it as a dictionary
instead of a string.

Test Description:
Copied and pasted the `csm-config` chart into a
`csm-config-manifest.yaml` file on mug. Downloaded the required helm
chart and container image onto mug and modified several values in the
manifest to isolate the impact of testing this change. Then shipped the
manifest with `lofstman ship --manifest-path `csm-config-manifest.yaml`.
Confirmed that the `loftsman` command succeeded, and the K8s job
succeeded as well. Inspected resulting content in VCS to ensure
`cray-sat` container version was set correctly.