CRAYSAT-1800: Update cray-sat to 3.25.9
Update the version of the cray-sat container image to 3.25.9 to include
the fix for the traceback in CRAYSAT-1800.

Test Description:
This version of the container image was tested on mug. The `sat showrev`
command was tested both with a valid and invalid filter, and `sat
bootprep run` was tested against an input file with multiple errors to
confirm that each line of a multi-line error message is prefixed with
the log level for compatibility with the IUF.