MTL-2393 Fix `goss-server` dependencies
Package `goss-servers` now requires `goss` to be provided (by any
package).
This was problematic in the NCN image build after
https://github.com/Cray-HPE/metal-provision/pull/675/files#diff-4b8036c5c83ba58cb56e769dfdde82765b1a451a424c8f93e114fa21775cc7f8
because `goss` was no longer explicitly called out, and aarch64 images
do not include hpe-csm-goss-package but rely on SUSE's own goss
package.

This new `Requires` will allow Zypper to auto-resolve which package to
pull in regardless of architecture.