CRAYSAT-1919: Update to cfs-config-util version 5.1.1
The cfs-config-util image is used by the `update-mgmt-ncn-cfs-config.sh`
script, which is used during patch installs. This script updates the CFS
configuration assigned to management nodes so that it uses the latest
version of the Ansible content from the csm-config-management repo in
VCS.

This new version of cfs-config-util adds support for CFS v3, and it
fixes a couple bugs that resulted in data being lost from configurations
that were modified with this tool (`additional_inventory` and
`special_parameters`). It also better handles unknown fields in CFS
configurations, so if, in the future, additional fields are added to CFS
configurations, they won't be lost.

Test Description:
Built a CSM release distribution that contained the shell script and the
latest version of the cfs-config-util image and copied it to drax.
Mocked out a new CSM release and associated branch in the
csm-config-management repo in VCS. Then executed the script to update
the configuration currently assigned to management nodes. The script
successfully updated 3 configurations assigned to those nodes and then
waited for the nodes to reach a configured status.

(cherry picked from commit 78c2d8af913d04b5983abd2511245c1dfb95265a)