Use LGI upgrade feature branch
We pushed a feature branch, feature/kubernetes-upgrade, to
livecd-gcp-infrastructure, to fix functionality in handoff.sh that was
causing K8s 1.32 vShasta instances to not start correctly. This
changeset updates the Jenkinsfile to use that feature branch for future
vShasta deployments for molly and orym as part of K8s 1.32/CSM 1.7
upgrade testing.