Cleanup disabling PSP for cray-sysmgmt-health
Remove explicitly setting "pspEnabled: false", the template
automatically detects when PSP is not supported, so it won't
add PSP stuff even if "pspEnables: true" is set.

Also, remove the pod annotation that is configuring AppArmor
in the chart, by setting it to "null".  It was setting it to
"unconfined", but we don't have any AppArmor profile confining
it, so it really wasn't doing anything.  And removing AppArmor
resolves the kyverno policy error due to policy exceptions not
working for AppArmor.