CASM-4937 Nightly Builds Phase: trigger nightly builds on release branches
Keeping integration and release branches in sync appears to be too activity
which often falls of the radar. We'll simplify nightly builds, by triggering
them directly on release branches. Thus, releae branch can be in "auto" mode
if it has "triggers {}" section in Jenkinsfile, and "manual" mode if
this section is disabled.