CASMPET-7443-CASMPET-7444 Fix generation of SBPS DNS records (#4063)
This updates the ansible code when generating DNS records
for iSCSI projection (SBPS), so that it won't include \r
characters to DNS, which causes DNS to fail.  This also
updates the curl commands so they will cause the script to
fail if the curl commands fail.

This also includes the fix for:

CASMTRIAGE-8069 DOCS and code: arp cache tuning topic has errors