CASMPET-7604 postgres CRD install/upgrade fix (#4211)
This pulls in a version of the cray-postgres-operator
chart, where the top level CRDs have been moved from
the /files directory to the /crds directory.  With this
change, and initial install will now use the CRDs in the
/crds directory, instead of the /charts/postgres-operator/crds
directory.  The difference between them is that the /crds
version enables patroni failsafe mode.